High-level control

High-level control.

spim_control(
  short_run,
  n_chains,
  date_restart = NULL,
  n_particles = 192,
  n_mcmc = 1500,
  burnin = 500,
  forecast_days = 57,
  workers = TRUE,
  n_threads = NULL
)

Arguments

short_run
Logical, indicating if this is a debug run. This will use freer particles, mcmc steps and a small sample

n_chains
The number of chains to run

date_restart
Optionally, dates save restart data in the pmcmc (see mcstate::pmcmc

n_particles
number of particles to be used in particle filter if short_run = FALSE

n_mcmc
number of steps to be used in PMCMC if short_run = FALSE

burnin
number of steps out of n_mcmc to be used as a burn-in in PMCMC if short_run = FALSE

forecast_days
number of days to forecast after the simulation

workers
Logical, indicating if we should enable workers. If TRUE, then a number of workers between 1 and 4 will be used depending on n_chains and the detected number of cores.

n_threads
Explicit number of threads, overriding detection by spim_control_cores

Value

A list of options

Details

TODO: document parallel detection