Load parameters

Load parameters for fitting.

spim_fit_pars_load(path, region, assumptions, kernel_scaling)

Arguments

path

The path to load from. This will be a directory containing files "info.csv", "prior.csv" and "proposal.csv" (for spim_pars_pmcmc_load) but also the baseline parameters ("base.rds") and a transformation function ("transform.R")

region

A single region, or vector of regions, to load.

assumptions

The name of any assumptions to apply to filter the baseline parmeters (e.g., "central")

kernel_scaling

The scaling coefficient for loading the proposal variance-covariance matrix.

Value

A list.

Details

There are lots of types of parameters here, depending on the perspective

This function deals with loading in the "logical" parameters; things like "vaccine efficacy" or "vaccination doses". These can be any R object.

Our aim is to create sircovid parameters, the result of sircovid::lancelot_parameters. This will be a list with structured parmeters that correspond directly to dust/odin inputs. Most likely these will be mcstate::multistage_parameters objects that cover multiple epochs with different model parameters, and these might be grouped together in a list so that a multiparameter (nested) filter can be run.

We'll end up working with small(ish) numeric vectors of parameters from the pmcmc process. Our transform function will convert this vector of parameters, using the logical parameters, into baseline parameters.

The mcmc parameters object that this function will return (as $mcmc) is a mcstate::pmcmc_parameters or mcstate::pmcmc_parameters_nested object, which contains all the information for the above.