Create parameters object

Create parameters object for use with mcstate::pmcmc. This function is very data-hungry as it (alongside spim_data()) is the main point at which data enters into the model. Here we end up setting a lot of data that are are not fitting to, but taking as fixed inputs (vaccination over time) as well as information on the parameters that the model will vary (their names, ranges, priors and their proposal distributions).

spim_pars(
  date,
  region,
  model_type,
  multistrain,
  beta_date,
  vaccination,
  parameters,
  kernel_scaling = 1,
  cross_immunity = NULL,
  waning_rate,
  sircovid_model = "carehomes"
)

Arguments

date

The date of the data. This will be the final date

region

The region that we are fitting to. This will either be one of the 10 NHS regions (one of sircovid::regions("all")), or an argument to that function (e.g., all)

model_type

Either "BB" (beta-binomial) or "NB" (negative binomial). We use and exclude different data types depending on this argument.

multistrain

Logical, indicating if the model is a "multistrain" model allowing for multiple competing strains.

beta_date

A vector of date (strings) for the beta parameters. Must align with parameters

vaccination

Vaccination data, from spim_vaccination_data

parameters

Parameter information, from spim_pars_pmcmc_load

kernel_scaling

Kernel scaling factor. Typically a number between 0 and 1 for scaling the proposal VCV

cross_immunity

Optional vector of cross immunity values. Only has an effect if multistrain is TRUE and then must have a length of 2 if given.

waning_rate

Rate of waning immunity

sircovid_model

The sircovid model, default is "carehomes"

Value

An mcstate::pmcmc_parameters object which can be used with mcstate::pmcmc