Prepare NPI key csv for simulation

Prepare csv with NPI keys for simulation task

spim_prepare_npi_key(
  schools,
  schools_modifier,
  country,
  path = NULL,
  gradual_start = NULL,
  gradual_end = NULL,
  gradual_steps = NULL,
  overwrite_central_adherence = NULL,
  overwrite_low_adherence = NULL,
  overwrite_high_adherence = NULL,
  npi_key = NULL
)

Arguments

schools
Should be one of 'open' or 'closed' and specifies which state values in the csv correspond to

schools_modifier
Amount to modify first strain Rt (excl. immunity). If schools = 'open' then the modifier will be subtracted from the given csv to create schools = "closed" scenarios, otherwise added - abs(schools_modifier) is applied internally

country
If "england" then all other nations filtered, otherwise no filtering.

path
File path to a csv containing columns: nation (at least one of england, scotland, wales, northern ireland), npi (name of NPI step for associated schedule), Rt (mean value for Rt), Rt_sd (standard deviation for Rt), adherence (low, central, or high)

gradual_start, gradual_end, gradual_steps
If all three of these are non-NULL then creates a scenario of gradual transition from Rt in scenario gradual_start to Rt in gradual_end in gradual_steps number of steps

overwrite_central_adherence
If non-NULL then a list with names corresponding to npi and values to overwrite the csv in the central/low/high adherence scenario. This may be be useful if automating some, but not all, of the Rt values

overwrite_low_adherence, overwrite_high_adherence
As for overwrite_central_adherence

npi_key
Optional data.frame, instead of reading from path

Value

tibble for passing to spim_prepare_rt_future