Prepare Rt future for simulation

Prepare Rt future grid for simulation task

spim_prepare_rt_future(
  npi_key,
  start_date,
  end_date,
  path = NULL,
  schedule = NULL
)

Arguments

npi_key
data.frame, rather than reading from path

start_date
Start date of simulation, all changes in schedule before this date are removed.

end_date
End date of simulation, all changes in schedule after this date are removed.

path
File path to a csv containing columns: nation (at least one of england, scotland, wales, northern ireland), scenario (scenario ID), year (YYYY), month (MM), (DD), npi (NPI ID to match npi_key$npi)

schedule
data.frame alternative to csv in path except with date column already create (and not year, month, day columns)

Value

tibble which is essentially a column binds of npi_key and Rt_schedule datasets after cleaning