Create scenario run grid

Create grid of scenarios to run for simulation

spim_run_grid(
  scenarios,
  csv = NULL,
  expand_grid = NULL,
  force_central = TRUE,
  set_strain_params = TRUE,
  multistrain = TRUE,
  analyses = NULL
)

Arguments

scenarios

Scenarios to run simulation over

csv

Path of csv to load run grid from

expand_grid

Optional large grid of scenarios such as from spim_expand_grid

force_central

If TRUE (default) then central analysis is always included as specified in simulation_central_analysis. This should rarely be FALSE as often required for basic checking plots.

set_strain_params

If TRUE automatically sets strain parameters strain_cross_immunity, strain_severity_modifier and waning_rate, as equivalent to strain_vaccine_efficacy

multistrain

If FALSE then removes all columns related to a second strain

analyses

If not NULL then filters grid by given analyses

Value

A grid of scenarios to run