Create control parameters

Create control parameters to run a set of simulations. This includes parameters that vary compared with the upstream simulation and grids of parameter types to loop over.

spim_simulate_control(
  flavour,
  regions,
  date_start,
  date_end,
  expected,
  parameters,
  grid,
  output
)

Arguments

flavour

The simulation flavour; this is a single string (e.g., "mtp" or "add_omicron")

regions

A vector of regions to simulate over. These must all be valid values in sircovid::regions

date_start

The start date, as an R Date object. Typically this is the last date of the data/fits.

date_end

The end date of the simulation as an R Date object

expected

A character vector of expected parameter names. These are the names of things that must be present within parameters, and this will be checked for you. It will be tempting to write names(parameters) here, but we suggest not doing that as this will provide a human readable description of what you are currently looking to simulate against.

parameters

A list of parameters. These might be direct replacements against the baseline or structured lists of parameters with names that correspond to those within grid. We'll document this more later!

grid

The parameter grid, indicating the set of simulations to run. It must have at least one row. Not all variables are allowed in here; in particular, we disallow most of the rt_ variables. An error will be thrown if anything unexpected is found.

output

Output control, created by spim_simulate_control