Create output control for simulations

Control output created as a result of running simulations. Many of the outputs are large, or require additional computation, so reducing this will make your life more pleasant.

spim_simulate_control_output(
  keep,
  time_series = TRUE,
  rt = TRUE,
  rt_weighted = rt,
  rt_type = NULL,
  state_by_age = FALSE,
  vaccination = TRUE
)

Arguments

keep

Character vector of outputs to keep. Must be names of variables as included in the index of the model run.

time_series

Logical, indicating if a time series should be output (these are quite large)

rt

Logical, indicating if Rt should be calculated. This is quite slow.

rt_weighted

Logical, indicating if weighted Rt should be calculated. This is also quite slow. Only allowed if rt is TRUE

rt_type

Character vecrtor of Rt types to use, passed to sircovid::lancelot_Rt. Can be any or all of eff_Rt_all, eff_Rt_general, Rt_all and Rt_general.

state_by_age

Logical, indicating if state by age should be output. These are really large and will cause memory issues for you if you try to run with long time series unless you have a lot of RAM.

vaccination

Logical, indicating if vaccination information should be output.

Value

An object of spim_simulate_control_output. Do not modify this object after creation.